feat: use Rackspace API to fetch credentials (#8)
# Description

The Python script using the Rackspace webpage is replaced by Curl which
uses the official Rackspace API. This API is a little bit faster than
the UI and easier to use. It is also used to retrieve a list of all AWS
accounts which are available in Rackspace for you. This list is cached
in `$HOME/.config/rackspace-aws-login/`.

The script asks for the Rackspace API token (check your Rackspace
account) and the Rackspace username.

# Verification

Manually executed and checked that I have access to the AWS account.

# Checklist

- [x] My code follows the style guidelines of the project
- [x] I have performed a self-review of my own code
- [x] I have made corresponding changes to the documentation