feat: retrieve Rackspace credentials from Keepass (#9)
Allows the script to access your local KeePass database and retrieve the
Rackspace credentials from there. To access the local database
`KPScript` is needed in your path and `KEEPASS_FILE` has to point to
your local KeePass database.

The database needs an entry `Rackspace` with the username and an
additional field `api-key`.