fix: only set aws profile with alias if alias is set (#30)
If no alias is given, the script would still configure an empty profile
block in .aws/credentials. This breaks the file, which is not good.
Also, some constellation of arguments caused local variables to be
accessed before set. To prevent that, all local variables are set to
empty string when created.